// src/types/kanban.ts
export interface KanbanStateProps {
  selectedItem: string | false;
}

export type KanbanColumn = {
  id: string;
  title: string;
  itemIds: string[];
};

export type KanbanComment = {
  id: string;
  comment: string;
  profileId: string;
};

export type KanbanSubTask = {
  id: string;
  title: string;
  completed: boolean;
};

export type KanbanItem = {
  assign?: string;
  attachments: [];
  commentIds?: string[];
  description: string;
  dueDate: Date;
  id: string;
  image: string | false;
  priority: 'low' | 'medium' | 'high';
  title: string;
  subTasks?: KanbanSubTask[];
  lastUpdated?: Date;
};

export type KanbanProfile = {
  id: string;
  name: string;
  avatar: string;
  time: string;
};

export type KanbanUserStory = {
  acceptance: string;
  assign?: string | null;
  columnId: string;
  commentIds?: string[];
  description: string;
  dueDate: Date;
  id: string;
  itemIds: string[];
  title: string;
  priority: string;
};